imported/exported function, and called it to display in console
