+++ src/app.js
@@ -1,9 +1,13 @@
+import getPokedex from './get-pokedex.js';
+
 // selecting the correct nodes
 const pokedex = document.getElementById('pokedex-info');
 
 // adding event listeners
 pokedex.addEventListener('submit', (event) => {
     event.preventDefault();
-    console.log('yo');
-    // const pokedexData = new FormData(pokedex);
+    const pokedexData = new FormData(pokedex);
+
+    const dex = getPokedex(pokedexData);
+    console.log(dex);
 });
