+++ test/tests.js
@@ -1,24 +1,6 @@
+import getPokedex from '../src/get-pokedex.js';
 const test = QUnit.test;
 
-function getPokedex(pokedexData) {
-    //convert data into variables
-    const hasBattle = pokedexData.get('battle') === 'yes';
-    const wins = parseInt(pokedexData.get('wins'));
-    const evolving = parseInt(pokedexData.get('evolving'));
-
-    //object variables
-    const applicant = {
-        name: pokedexData.get('name'),
-        nickname: pokedexData.get('nickname'),
-        primary: pokedexData.get('primary'),
-        secondary: pokedexData.getAll('secondary'),
-        hasBattle: hasBattle,
-        wins: wins,
-        evolving: evolving 
-    };
-
-    return applicant;
-}
 test('create applicant from data', function(assert) {
     //Arrange
     // Set up your parameters and expectations
